backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
base58==2.1.1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
deprecation==2.1.0
exceptiongroup==1.2.2
fastapi==0.115.0
firebase-admin==6.5.0
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gotrue==2.8.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
ipfsapi==0.4.4
ipfshttpclient==0.4.13.2
msgpack==1.1.0
multiaddr==0.0.9
multidict==6.1.0
netaddr==1.3.0
opencage==3.0.2
packaging==24.1
postgrest==0.16.11
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
realtime==2.0.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
tqdm==4.66.6
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
varint==1.0.2
websockets==12.0
yarl==1.12.1
